PycharmProjects/python final/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.35
psycopg[binary]==3.2.12
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
